# security: prevent domain hijacking by using .fastn TLD for email addresses